requirements.txt
chardet
requests
urllib3
python-nmap